test: unsigned 제거

+++ test/cipher_net/step2-2/server.cpp
@@ -71,7 +71,7 @@ int main(void)
 
 	// timestamp의 마지막 12문자를 uint8_t로 변환 후 nonce2에 복사
 	std::transform(timestamp.end() - 12, timestamp.end(), nonce2.begin(),
-				[](char c) { return static_cast<uint8_t>(c); });
+				[](char c) { return static_cast<int>(c); });
 	std::cout << "after nonce2 size: " << nonce2.size() << std::endl;
 	// nonce 값 출력
     std::cout << "Nonce2: ";
